requirements.txt
discord.py
pyfiglet
gdshortener